\name{LogConcDEAD-internal}\title{Internal log-concave maximum
  likelihood estimation functions}

\alias{initialf}
\alias{initialy}
\alias{convhullnew}
\alias{delaunaynew}
\alias{integratemarg}
\alias{JAD}
\alias{JAD.approx}
\alias{JAD.ord}
\alias{logliklcdmix}
\alias{dslcd_1D}
\alias{dslcd_2D}
\alias{dslcd_com}
\alias{simplex_unit_05_2_nd_pt}
\description{
  Internal functions for \pkg{LogConcDEAD}
}
\details{
  These are not to be called by the user.
}

\keyword{ internal }
